Fixed type conversion

+++ src/cpp/cpp.cpp
@@ -50,7 +50,7 @@ bool cpp::write(const vsg::Object* object, const vsg::Path& filename, vsg::ref_p
     {
         // long string has to be handled as a byte array as VisualStudio can't handle long strings.
         fout << "uint8_t data[] = {\n";
-        fout << static_cast<uint32_t>(s[0]);
+        fout << uint32_t(uint8_t(s[0]));
         for(size_t i = 1; i < s.size(); ++i)
         {
             if ((i % 32)==0) fout << ",\n";
